backend/foodgram/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2.15
django-cors-headers==4.0.0
django-extensions==3.2.1
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.4.0
exceptiongroup==1.1.1
flake8==5.0.4
idna==3.4
importlib-metadata==4.12.0
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.7.0
pytest==7.3.1
pytest-django==4.5.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2023.3
reportlab==3.6.10
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
zipp==3.8.1
gunicorn==20.1.0
